:sparkles: 7장 1차 정리 완료 kafka 관련부분은 수행결과 확인하지 못함

+++ ch07-Reactor/microservices/review-service/src/main/java/se/magnus/microservices/core/review/services/ReviewServiceImpl.java
@@ -38,37 +38,25 @@ public class ReviewServiceImpl implements ReviewService {
         this.serviceUtil = serviceUtil;
     }
     
+//ch07
     @Override
-    public List<Review> getReviews(int productId) {
-
-        if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
-
-//        if (productId == 213) {
-//            LOG.debug("No reviews found for productId: {}", productId);
-//            return  new ArrayList<>();
-//        }
-
-//        List<Review> list = new ArrayList<>();
-//        list.add(new Review(productId, 1, "Author 1", "Subject 1", "Content 1", serviceUtil.getServiceAddress()));
-//        list.add(new Review(productId, 2, "Author 2", "Subject 2", "Content 2", serviceUtil.getServiceAddress()));
-//        list.add(new Review(productId, 3, "Author 3", "Subject 3", "Content 3", serviceUtil.getServiceAddress()));
-
-        if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
-
-        List<ReviewEntity> entityList = repository.findByProductId(productId);
-        List<Review> list = mapper.entityListToApiList(entityList);
-        list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
-
-        LOG.debug("getReviews: response size: {}", list.size());
-
-        return list;
-    }
-    @Override
+//    public List<Review> getReviews(int productId) {
+//        if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
+//
+//        List<ReviewEntity> entityList = repository.findByProductId(productId);
+//        List<Review> list = mapper.entityListToApiList(entityList);
+//        list.forEach(e -> e.setServiceAddress(serviceUtil.getServiceAddress()));
+//
+//        LOG.debug("getReviews: response size: {}", list.size());
+//
+//        return list;
+//    }
     public Flux<Review> getReviews(int productId) {
+
         if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
 
         return asyncFlux(getByProductId(productId)).log();
-    }
+    }    
 
     private <T> Flux<T> asyncFlux(Iterable<T> iterable) {
         return Flux.fromIterable(iterable).publishOn(scheduler);
@@ -101,7 +89,9 @@ public class ReviewServiceImpl implements ReviewService {
     
     @Override
     public void deleteReviews(int productId) {
-        LOG.debug("deleteReviews: tries to delete reviews for the product with productId: {}", productId);
+        if (productId < 1) throw new InvalidInputException("Invalid productId: " + productId);
+        
+    	LOG.debug("deleteReviews: tries to delete reviews for the product with productId: {}", productId);
         repository.deleteAll(repository.findByProductId(productId));
     }
 }
